unit_test: Add simple test to see that random, custom timeperiods work
This covers 43410fe, to make reports not crash on timeperiod exception
overlap, and it also uncovered a bug in ocimp, so I hereby declare the
time spent writing this test well-invested.

It still couldn't care less about whether the result is correct, but it
does catch crashes and overly bad things.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>