Makefile: Give test config environment dirs rights 777
We're often going to be running this as one user, often build but when
testing changes I'm often root, that is different from the monitor user,
usually monitor. If monitor can't write it's state to the var dir, we
can't get ocimp to import the state files as they don't exist. This
creates tears and sadness all around. So fix by making the directories
super-permissive.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>