host model: Make get_hosts_for_hostgroup a wrapper to hostgroup model
Turns out, the exact same method exists twice. One of them executes
quickly (the one in the hostgroup model), the other doesn't (the one in
the host model). Instead of trying to keep both executing quickly,
simply make one a wrapper to the other. Let's keep the fast one around.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>