getparams helper: Fix so that it does something remotely sane
Not sure exactly what it was doing, but it sure wasn't correct if there
was more than one query parameter.

Fortunately, we have tests checking that specific case.

Unfortunately, nobody bothered to have a look at them.

This fix is already in master as of d2d2614, but as that commit contains
other stuff I don't want here.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>