user import: Move user "expiration" to common code
This fixes the (completely, afaics, harmless) bug, where LDAP users
weren't getting expired. This meant they were always kept in DB, even if
they were no longer authorized for anything, but this should be
harmless since there are authorization and authentication checks
executed on the users anyway.

This also fixes the more serious problem with fallback auth, where the
following works:
 $config['driver'] = array('Ninja'=>'Ninja','LDAP'=>'LDAP');
but the following doesn't:
 $config['driver'] = array('LDAP'=>'LDAP','Ninja'=>'Ninja');
The latter is what the manual recommends, so that's what people does.

It breaks because ninja auth tries to clean up deleted users, and since
it doesn't understand the LDAP users, it deletes all of them. Ouch. By
moving this to common code, we collect all backends' users before
deleting anything.

We also make sure we only run the htpasswd import once (but, yes, we
parse the htpasswd file loads of times), and that it's executed after
the LDAP import, so the LDAP import can empty the password field, and
htpasswd can set it again - without this, deleting the monitor user from
the htpasswd file wouldn't have any effect if there was a monitor user
in LDAP.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>