nagstat helper: Fix us and euro date formats
They were previously listed with hyphens instead of slashes as
separators, which doesn't fly well with PHP's strtotime()
function.

Why anyone in the entire world would want to use anything but
iso8601 format date and time is completely beyond me, but since
we claim we support such follies, we might as well do it right.
This patch makes sure we do, even if it is an atrocity to begin
with.

Signed-off-by: Andreas Ericsson <ae@op5.se>