reports: remove all alert summary and/ors for unlimited users
When we ask for an alert summary for *all* hosts and/or for *all*
services, which is likely the common case due to it being among the
"standard" choices, and we're an unlimited user, which is likely the
common case due to unlimited users being the types of people who are
interested in network-wide information, we don't need to ask for,
generate, and send in a filter on *every* host and *every* service in
the database. Also, we don't need to filter on event types twice.

This should make things a teeny, tiny bit faster, and looks less insane
when you examine the SQL queries. This might mean this feature works for
users with a few more hosts before triggering bug #5144. It's not
actually *fast*, due to us still asking for all rows in the database,
and once we've downloaded them, in PHP, skipping the lines that doesn't
represent state changes, which is silly when we could just use count(),
but required due to bug #5056.

Seems to work as a random unlimited user and a random limited user for
"All{ host, service,} state changes". Someone, please, cane me until I do proper tests of
this patch sober.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>