ninja controller: Anal isset()
If you use isset on something that does half a dozen nested array
lookups, it won't actually protect you from the case when one of the
keys might not be set in one of the inner arrays. Adding a few more
isset cases made ninja not crash for me on some random occasion when I
did something stupid, and things that prevents crashes are things I as a
rule like.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>