Tests: Initialize unit test library before running test
We want to override error and exception handlers so that test runs fail
if there are errors or exceptions, and we can't do that when we're
already in the view. Thus, do it from the controller, and as a
consequence, make unittest::instance() a singleton - anyway, that's what
the name suggests that it already is.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>