report model: Set st_needs_log more rarely
It needs to be false in order to allow generating large reports, but it
needs to be true for the trends graph. Thus, remove the previous
unconditional lines setting the log to true way too often, and instead
only set it to true if we include trends. This once again allows us to
generate massive maps without scary large memory use.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>