upgradescripts: Try to convert the order string when upgrading
I'm feeling this might break horribly in weird, random cases that I
don't understand, but it does handle the normal case quite alright, so I
guess that's better than not handling it at all, as was the case before
this change. Should this break, the user's default widget order will be
weird until (s)he's fixed it again, so it's not a disaster.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>