host model: Be less paranoid about user rights
A user that is authorized for a service should be allowed to see the
service, even if they're not authorized for the host the service resides
on.

A user authorized_for_all_services can still not, with this patch, list
services (which they probably should be, but bleh), thus this bug wasn't
very obvious when testing ninja, but the user can quite correctly see
the services in our business processes for example, and from there go to
the extinfo page, so the extinfo page must not crash in this case.

This fixes bug #5281.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>