ajax: Make widget settings have 1 change event handler always
It used to be that they had 0 if they were added after page load. By
adding the event handler on widget init, I got that up to $n, where $n
is the number of times the widget has been loaded. After doing a
poor-man's "unique" on the javascript register_widget_load array, I got
it back down to 1, always.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>