js: Make toggle_label_weight() smarter
Previously, you had to put an ID on the input element's label. Now, just
give the function the ID you already got (of the input element itself).

The old approach should be cleansed and all obtrusive calls should be
replaced with a hook attached to the input element's not yet existing
.js-fatten-my-label class.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>