widgets: Fix set global widget refresh rate input event handler
When selecting the input element in the global "Set widget refresh rate"
with the mouse, you click, causing a click event, which wasn't handled
by the input, and therefore bubbled up to the parent div and hid away
the input. Probably not what the user was expecting.

Fix by only doing a hide if the event target is actually the div itself.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>