alertlog: Don't join when filtering on {host,service}groups
This sounds like a terrible decision - nested queries are bad, right? -
but since the inner query only returns a bounded number of lines (1000
services in one service group is quite a lot), and the outer query often
has a ginormous number of rows (that are mostly discarded and
count():ed), creating a temporary table with a whole lot more columns
than a plain report_table is stupid - especially when we don't want
those columns.

When filtering on a large-ish service group on a test system, the
old count() query took about 1 minute, while the new one takes 10
seconds. By also installing an event_type/host_name/service_description
index, we can get that down to just somewhat over 1 second. This was on
a fairly small system with only half a million lines of report_data,
which would indicate that the performance improvement might be even
bigger in practice.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>