outages: Calculate correctly, once
We used to do it 3 times, all of them incorrectly.

We failed to include the services on this host when seeing how many
services this outage affected - but we did include this host in the
number of affected hosts.

There's was a bunch of code to get every non-OK hosts, and then filter
them so that only the parents will show up. This seemed buggy,
unpredictable, and in particular ineffective. This has been replaced
with a simple check if the host is down or unreachable - finding the
source of the outage is, after all, tho whole point of having those two
nagios states in the first place.

This should take care of bug #5325.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>