reports: Make top_alert_producers slightly less useless
Because we tried to iterate over the result set, kohana's database layer
decided to do the most moronic thing it could conceivably do: do an
infinite number of database seeks, or as many as it could do before it
hit PHP's max execution time and was killed.

To quickly work around this, we cache the whole result set in RAM. This
is stupid, but it's less stupid to use lots of RAM than to do nothing
useful at all.

Some day, someone will see this change, scream, and do something good
instead, while banging their head at the moron who thought "cache
everything in local ram was a good idea". That will be a good day.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>