reports.php: Handle combined reports like UP-unscheduled/DOWN-scheduled
When one host/service was UP/OK without scheduled downtime, and one
host/service was DOWN/WARNING/CRITICAL in scheduled downtime, we'd say
that the worst state is DOWN/WARNING/CRITICAL outside of scheduled
downtime, which is a lie.

This becomes even murkier when one service is WARNING outside of
scheduled downtime, and one service is CRITICAL in scheduled downtime.
The worst state, then, is reasonably dependent on whether we count
scheduled down as up or not, because if we do, WARNING is worse than
CRITICAL, but not otherwise.

There are some new comments trying to sched some light on this, and -
more importantly - there are a set of test cases.

This patch adds some bookkeeping overhead to the state calculation, but
hopefully it shouldn't take too much resources.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>