view/command/request: Make sure element is set before using
This is a fix for commit e703759. It made us use the name instead of
some internal ID when scheduling downtime for a set of objects.
Unfortunately, in the Real World, it's possible to have the default
array reference stuff not in the options array, causing crashes. Simply
falling back to using numeric IDs in exceptional circumstances sounds
like a sane fix.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>