css: Add one single "bottom: 0px" to menu css
...and a "height: 100%" to the template.

Also remove a bunch of fucktarded js that never worked and meant that
every second time you viewed ninja on a less-than-full-HD monitor, you
couldn't press "configure" because you couldn't scroll because the js
that controlled the style was fucktarded and never worked.

So, don't spend a lot of time and effort on intentionally ruining the
browser's idea of what to do and doing a slow, poor job doing it
ourselves. Instead just do nothing, and everything works like Tim
intended.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>