themes: Make common.css common
Do this by introducing a new screen.css, and keep only the difference
between print.css and screen.css in common.css.

Now, the differences are highly erratic and inconsistent - print.css is
an old copy of a slightly buggier version of the old common.css (now
screen.css), so the differences were in many cases due to screen bugs.
However, now it's at least possible to find the print CSS code that's
actually for print.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>