Format dates according to Nagios date_format variable
This patch replaces hardcoded date format strings with calls to
nagstat::date_format(), resulting in dates being displayed according to
the date_format variable from the Nagios configuration. The helper
method caches the config value in the database as a setting, making it
reasonably trivial to implement a per-user date format setting in the
future.