models/current_status: Rewrite to use livestatus instead of merlin db
The two data sources doesn't completely agree on what specifies a
pending check, but this seems to be a merlin bug, as merlin seems to
randomly set pending checks to OK since 1970-01-01.

Other than that, this seems to work just as well, though faster. How
much faster? Well, that depends.

In order to simulate a widget refresh, when all widgets simultaneously
attack the server asking for new content, I made current_status not
cache the results, and used Kohana's built in benchmarking tool to get
some numbers.

If the config is small, this doesn't really do much of a difference, as
both methods are pretty much instant since mysql can keep everything in
RAM and be quick. However, as I grew the config to a few tens of
thousands of objects, a livestatus backed page load took 20% of the time
that a mysql backed page load - and this is just one changed model, it's
possible that the remaining 20% are still primarily spent by mysql, but
in other models. These numbers seem true regardless of whether the user
is unlimited or not.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>