benchmark: Log to file
Looks for $config['log_benchmark_to_file'] and tries to write benchmark
data to that file, a space separated list of timestamp, user, url,
execution_time and memory_usage.

That config setting should *not* be used at all times but rather kept on
for a small period of time to collect a debuggable log file.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>