Remove a ton of get_authorized_{hosts,services}
These are evil. They take a ton of time, and they're almost never what
we actually want - especially when the question is "can we see this",
asking "give me all objects I can see so I can see for myself" is, well,
excessive.

On my machine, with a large config, this gave me more than a 10-fold
load time improvement - from 15-20 seconds down to 1.2 or so - on the
service overview. That's pretty nifty.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>