Nagios_auth_Model: Turn into a singleton
When experimenting, I noticed that get_authorized_hostgroups is called
32 times for every hostgroup overview page load. I did not check exactly
how many of them were cached due to sharing auth instance, but looking
at the code I'm going to bet "not that many".

By using the model as a singleton, all it's results are cached. We still
allow instantiation to not break integration projects, though we look at
anyone who does instantiate it with a frown.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>