helpers/ninja: Retrieve registry instance only when we need to
This one, small, tiny fix have *massive* effects. Really massive. Doing
something like
  time php index.php status/hostgroup monitor > /dev/null
(so the terminal speed is not a factor) on a large-ish config, this
reduces the execution time by 86%. I didn't mistype that - we spent 14%
of our time doing things, 86% of our time waiting for zend to return a
stupid singleton instance. Which we then ignored. Jay us.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>