Remove FirePHP
The FirePHP profiler looks cooler than the vanilla Kohana one, but it
isn't any more informative, it's Firefox-only (which I'm pretty much the
only dev that uses), and if you're unlucky (which I am) it breaks and
you have to delete your Firefox profile and it might start to work
again. So, now I kill it.

While we're at it, leave the decision whether to use the profiler up to
the profiler config, stop doing stupid and incorrect things to the weird
ALLOW_PRODUCTION flag.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>