views/group_overview: Don't use kohana helpers
They create a *ton* of overhead.

This only affects one template, and should thus be seen as a PoC, but it
shows that for pages built around that one template, not calling these
reduces rendering time by a full third in my tests with a large config
(and a fairly warm mysql).

So, if somebody has a day extra, I know what they can do...

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>