livestatus: Use database config, include in benchmarks
It *is* a database, albeit a weird one, so let's make it's input and
output look more like a database is supposed to look in Kohana.

Because it doesn't use SQL, I suspect it would be more confusing than
helpful to actually base it off the Kohana database library, though.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>