models/config: Implement the count option properly
When trying to find out the number of pages to use in the pager, instead
of retrieving all data on all objects in the config and counting them,
only retrieve a count(). This is faster.

How much faster? On my test config, it brings down the config rendering
time from the minute range to the sub-second range. So "quite a bit".

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>