*group: Optimize the "ask for more groups than available" case
Cases have been observed where this gives us a 5 fold performance
improvement for the service query (which is the slow query) when asking
for a gazillion hostgroups and there "only" are a few hundred available.

I am simultaneously impressed and grossed out by myself for comming up
with this method to achieve acceptable performance in yet another one of
the cases that we benchmark.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>