widgets/status_totals: Use stats model instead of status_totals one
The status_totals model retrieves a ton of data that the status_totals
widget doesn't care for. The singleton-ness would potentially give us a
performance benefit, except we only use the status_totals widget on
pages that never, ever otherwise creates an instance of the status_totals
model, so this should make these pages a teeny, tiny bit snappier.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>