config: Fix multi-val columns
We assumed there was never more than one {parent,contactgroup,contact}
per {host,service}, which is an interesting assumption, since there
would be no reason to create a join table then. This fixes bug #5195.

We also did implode(', ', explode(',', implode(',', $result))), which
seems silly when you look at it.

Finally, we completely failed at joining together the contact and the
contactgroup objects in that there was always an extra comma if a
host/service didn't have both of those.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>