refresh: Half the number of regular ajax requests
Because round-trip time to a tired server can be huge, and the whole
async-ness of browsers is highly exaggerated, we shouldn't have too
many.

Use our json helpers to generate "standard" output, so we can easily
detect error conditions in the response, and have the "Updated" field in
ninja mean "Updated" rather than "current time" - when things break, it
will show when it broke. While we're at it, do add the ugly yellow bar
when the server isn't responding at all - that might actually be sorta
serious.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>