oracle driver: Add the ugliest hack yet
Oracle doesn't support sorting by CLOBs, "output" is one of those, and
we do support sorting by it. What to do? Well, "output" is always a
CLOB, and when we order by it, we can str_replace it with a
clob-to-varchar method wrapper.

So sorry about this.

This closes #5539, again.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>