controllers/command: Make it possible to reschedule checks
So...

At first there's a line of code ensuring $host_names is an array, even
when it isn't, with a comment explaining why we need to do so.

Then, a dozen lines later, there's a comment suggesting that $host_names
probably can't be an array, and code relying on that. Yeah, right,
that's gonna work.

So, who's the moron that wrote that code? Oh. Right. Me. Of course.

Well, at least it evolved into a broken state over the course of time,
so that it at first worked, before another fix broke the first case? Ah,
all of it is from c06e284a.

I'll just go home now, for everyone's sake.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>