Move SQL files to sql folder
This means they're in a place of their own, looking neat and tidy,
with mysql and oracle separated from each other, and being all
consistent with nacoma and merlin.

Also remove all verbose prompt messages from ninja_seq.sql, and rename
it ninja-install.sql to reflect that it's only meant to be executed
once.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>