install scripts: Cleanups
Kill "merlin-reports-db-upgrade" and merge it with the regular
ninja_db_init - half the former file was copy-pasted from the latter,
which is usually a sign of high code quality - however, CLI args weren't
passed along, so that's something that's different.

Killing it also means we no longer break when we try to run this in
non-RH environments - for some "interesting" reason, we made sure to run
reports-db-upgrade with "sh", even though it was marked executable and
the shebang in it clearly said bash.

Remove the ridiculous $prefix thingy - we always expect this script to
be located in $prefix already, so there's no reason to force the user to
say that the directory they're running it from is the directory they're
running it from.

Copy-paste some arg handling code from merlin (...because that's a sign
of code quality :/ ) to add a more sane way to specify database
connection settings without having to modify the script.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>