config: Search through all result pages
This modifies the filterbox to be included in the page reload and thus
no longer be already-in-DOM-filter only. The searching is only
implemented for the primary keys which should be feasible enough, if
not, the code is easy to adapt for multicolumn filtering.

Also, some cleaning was done which affects largely the same rows/files
as the big diffs.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>