Close XSS in check outputs
Any user that is authorized to submit check results could submit one
that injected stored XSS into the page. In particular the alert history
makes this quite nasty, as it tends to stick around on that page.

Fix by HTML-escaping the output and the perfdata.

This is related to bug #5871.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>