reports: Move all repetetive date-related report code to reports helper
We usually don't use it - especially since the commit I don't know
where it ended up that fixes how SLA reports does things so it doesn't
use month names in arrays no more.

The only time we do use it is in javascript, so keep it specific to
javascript generating code.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>