reports: Make handling of start/end times less haphazard
We had what felt like hundreds of variables, all dedicated to the task
of storing the value of the current date format. Have the templates ask
the options object for proper formating (which itself might be a
ridiculous API, but it's an API at all) when they need it, and we can
delete even more code.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>