reports: Remove "assume initial" options
When importing from logs, we never, ever write initial states that are
the same as the last state prior to shutdown.

The fact that the merlin module does do that is considered a bug - the
much talked-about, highly-prioritized bug #5056 in fact - which means
that as soon as that's fixed, we will /never/ write states that are the
same as before shutdown to the database.

Because we, in this situation, can never be sure when the initial state is
over and the old state returned via a check result (if it is), we can't
work with an assumed state other than the last one prior to shutdown, or
the results will be... "Undefined".

While admins are notorious for their obsession to tweak and fiddle with
stuff, having a "broken/not broken" toggle is really not helpful at all.
Just remove the feature completely.

Thus we as of now always assume use_retained_program_state is activated,
and we use the same way to count time as nagios does in this situation.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>