reports: Always pass object list as arrays
This means we always generate a sub-report in the report module. This
makes the simple case (=1 host/service) somewhat trickier - in fact, it
becomes just as tricky as the common case (>1 host/service), except
there aren't a dozen if statements to decide which case it is, which
we always gets wrong, and our tests almost always execute the simple
code path, and our users almost always execute the other.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>