reports: Fix for when reports start when nagios is off
We used to call this "DOWN but <last state>", even if we weren't
assuming states during not running. For released versions, we still
handled this in our avail table (mostly), but the log and trends graphs
looked fishy.

For this rework branch, ninja never figured out that nagios had started
(the sub report did, but not the parent), and this caused bugs where
tests wolud return 14000 seconds shut down for a 9000 second report.

Now, unless assumestatesduringnotrunning, use -1 as state when initial
state involves a shut down nagios. We must still record the last real
state appropriately, though, as we'll use it as soon as nagios starts.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>