Fix custom page headers and footers
This works around wkhtmltopdf bug #957. While it's now possible to
override these to set your own (per installation) header/footer, it's
not recommended. As soon as webkit's support for paged media improves -
which is being worked on right now - you'll be able to set this
per-skin, which will be a much better solution.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>