models/reports: Remove fallback, auth for alert_summary query
The alert summary query generator validated the authorization for any
data returned, which is normally good, except when we already know we're
authorized to see this data, in which case it - especially when done
though the "give me all the objects I can see so I can see if this one
is part of that list" method - only wastes time.

There was also a fallback for object selection - when we chose to
include host events, without choosing any actual hosts, it was assumed
we really meant to say "all of them". Now, this is not only slow, it's
also utterly unhelpful.

This means we actually fix bug #5637.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>