schedule: Generate schedule table in js
We need to do be able to do so anyway, since we're able to add
schedules via ajax.

We don't want to be updating both the JS and the view and assume people
will remember both - I synced them up two days ago, and yesterday
someone "de-synced" it again.

So, do everything in JS always. In fact, put a template in HTML, so we
don't have to generate all the elements, but only copy and edit it
somewhat.

In theory, this means anything that ever works will always work. In
practice, well, it should still break less.

This also makes all the icons all consistent, instead of one being
"a > img", one being "div > img", and one being "form > input" (which is
my favorite). They position themselves, they line up, and they only have
to be specified in CSS once. Aaaah.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>