reports: Change how state filtering works
State filtering now requires you to tick the checkbox to *remove* the
state, rather than ticking it to *include* the state. This means the
ticked states create a mapping between from-state and to-state. By doing
this, we can make SLA reports map any excluded state to OK/UP, while
avail reports keeps mapping the state to the magical "excluded".

This resolves bug #4401.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>