Report_options: Set default report period to last7days
It was set to "false", which resulted in "today" in practice. Because
all report types either override it (SLA works on months, Avail has
last7days) or are horribly annoying because of it (Summary), it probably
wasn't a very good choice. Set it to last7days here already, and summary
will pick it up as well. Consistency!

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>