summary: Fix Some Bugs™
Because js addition was spread out throughout the method in different,
interesting ways, it was very hard to figure out which files and method
calls were missing from the generate method that caused the javascript
calendars to fail miserably, while they worked on the index page.

After moving the code about a bit, a couple of interesting lines were
added, but the moving about was kept around to simplify this type of
debugging the next time.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>