summary/latest: Fix rendering of comments
Because there was a nowrap, everything just kept going on and on
forever, which looks bad on screen, but looks extra bad when your PDF
renderer zooms out to allow the full page to fit on the page. Grrrr!

Remove the nobreak, and also use an actual table to emulate a table - it
works better, and is probably much better supported.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>