views: Show right click menu when it's supposed to be there
Before the livestatus port, we only showed it in totally weird
conditions that basically came down to "you need EVERYTHING" - so I
guess it only worked reliably if you were testing the default config as
the monitor user.

After it, we needed authorized_for_all_{hosts,services}, which still
makes no sense, as we can still be authorized for commands.

From now on, use the logic we use to detect if we're authorized
for submitting the command, tosee if we should show the menu.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>