libraries/Livestatus: Don't add more slashes
The config - which is in use - already has three slashes in the path,
so adding even two further? That's where I draw my limit.

More importantly, that's where fsockopen draws its limit as well, so
rather than opening the socket, it doesn't. Now that ninja depends on
livesocket for pretty much everything, that's double-plus ungood.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>