Restore "pending" state
We used status_text in models/current_status to find out the name of our
current state, based on our state code. In livestatus, the state code is
never 6 (aka pending), because we made that code up ourselves. We must
instead check if it has_been_checked, which means an extra arg (because
the name of those columns is sometimes "host_" prefixed, so we can't
just send in a row object).

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>