nagios_auth: Implement "is_authorized_for"
It's pretty easy to do, now that our livestatus lib takes care of adding
AuthUser for us - if we can read the name, we can read the row.

Because I'm afraid of some piece of code trying to use merlin integers
with these, I have a temporary exception in place to crash such code.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>