reports: Make custom report periods less broken
It was impossible to edit custom periods in summary reports, and it was
impossible to save custom periods in any report. It was also impossible
to load custom time periods in some places.

It's basically resolved by doing less. Don't update already set values
when loading, don't update already set values when saving.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>