tests/reports: Make tests explicitly include soft states
Some of them did so implicitly, as that was the default in the model,
yet not in the UI, which is why things worked out the way they should in
the past, but when I refactored the reports to make the Report_options
control the defaults of both, stuff became out of sync.

Bleh.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>