system/Event: Make event name accessible
It used to be abso-stinking-lutely impossible to access the event name.
It is now available, in the same fucktarded way that the "data"
parameter is - meaning it's just as broken as everything else, but no
more so, and actually fixing it requires major surgery.

Seriously, how do you even come up with an event system this stupid?

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>