Add method for exporting auth
It's connected to ninja_db version 4, which is only slightly wrong, but
adding it to post in the spec file would be worse, as any lingering auth
stuff would get annoyingly and constantly reexported. It does mean that
any current users of v2.1-beta will need to manually run this.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>