template_header: Don't choke when logged out
Since ninja's whole controller inheritance combined with "do anything in
the php object's constructors and construct 5 levels of inherited
objects before we decide to take action on non existing sessions", we
end up with generating the logged-in's layout and throwing it away,
only using the default logged-out's theme. Poor design.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>