external_widget: Allow user to specify group
Because of the auth system, it's not possible to fetch any user's group
effortlessly, because it might be an LDAP user, in which case we'd need
a password, and not asking for a password was sort-of the whole point.

So, we'll add a "feature" that allow the user to specify an array of
groups, to make it work like it did before.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>