Remove old geomap widget
It's a remnant of our old nagvis integration, back when we decided a
suitable name for the nagvis widget would be "geomap" (it wasn't as
illogical back then as it sounds now, hysterical reasons, etc). We
merged the 1.5 nagvis integration into ninja proper already, so get rid
of this particular piece of brokenness.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>