reports: Send reports with boolean options
The significant diffs are:

- report_controller: Set auth for php request via CLI
- schedule_controller: Send a report id instead of options, in order to
  be able to...
- report_options: React to saved reports, in browser or via cli (the
  extra output_parameter), to enable fetching of boolean values such as
  include_trends.

Before this, we were not able to send reports including boolean
settings.

As a note, if you try to debug reports, do *not* use Google Chrome's
built-in PDF-viewer since the content appears to be scrambled but once
downloaded, the PDF report is displayed correctly (eh, at least in OSX'
Preview).

Signed-off-by: Carl Helmertz <chelmertz@op5.com>