bootstrap: Skip Kohana if you want to
This makes sure that we can include Ninja's index.php in, for example,
right out of the blue; the HTTP API, simply by typing
define('SKIP_KOHANA', 1) before requiring the index file.

It will disable the routing but let you use the (completely insane but
relied upon) autoloader of Kohana's.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>