controllers/backup: Fix reload commands
It had been copy-pasted from the restore command, so we concatenated
strings from things like the backup filename. Right before that, we had
the "touch random file" command, which I removed in the parent commit,
which made the backup try to touch the tarfile. It's not allowed to do
that, though, so that generated an error. Then, the test verified that
we did in fact return an error, and if so, displayed "Restore
successful".

Now I've spent a couple of days fixing this, because it always looks
like "just one more thing", but this time, this should be it. I have
high hopes. I mean, I have hopes. You know, hope. A small hope. A bit of
hope. A tiny shrivel. I mean, I'm sure this is gonna fail just as hard
as everything else, but I know where to acquire alcohol.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>