controllers/command: Fix command rights
We used to say we were always authorized for all host/service group
commands (now we're not), we used to say we were authorized for command
if the first object did (now we do not), we used to pretend that any
command that has both a host name and a service description set was a
host command (now we do not), we used to look at an incorrect argument
variable to find service names (now we do not), we used to ensure that
we were authorized for host commands when submitting service commands
(now we do not), and we used to crash if a limited user tried to submit
the command (now we do not).

This fixes bug #6349.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>