Fix report options for SLA reports in op5reports
Due to unhealthy circumstances, SLA reports refused to use the provided
report period option, which caused the default value to always be used
instead.

As today is Jan 7, only ever generating SLA reports for "this year" is
a somewhat unhelpful default value, but it only becomes thoroughly
annoying when it turns out it's impossible to change.

Fix by convincing Report_options that "last3months" is a legal value,
and then remove it again in Avail_options, thus nothing should really
change.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>