form helper: Be less fucktardedly broken in order not to kill IE
So, here's a nice quiz for you!
<select id="flubb" option="Array">
	<option value="first">First</option>
	<option value="second">Second</option>
</select>

$('#flubb').val()

What's returned?

On non-IE: the selected option.
On IE: either "second" or "" depending on which element you select.

Now, in IE's defence, while it's being - how do I put it nicely? -
incredibly stupid (unless it's just being stupid, and it's jquery that
supplies the incredibility), Kohana is being utterly moronic first.
Because IE actually works when Kohana doesn't invent new HTML attributes
which are only there to act as decoys.

That was fun to discover - now I'm going to have to spend all weekend
drunk out of my mind just to process this.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>